This change fixes #5294.
The live development didn't work if the path for the --user-data-dir (passed to Chrome(-ium) commandline didn't exist. Previously a path to /tmp/ ws constructed, but some users reported that this didn't work on their systems.
In order to fix this behavior, we now use the appSupportDirectory to store the user data directory.

The new location for the user-data-dir is now in ~/.config/Brackets/live-dev-profile.
To determine the location of the appSupportDirectory, the call to gtk_init had to be the first when the app starts. This will ensure that g_get_prgname() and g_get_config_dir() will return valid locations.