Proper fix for "Show in OS" for Linux.
Now using a IPC call to FileManager1, similar to what Firefox does.
The fallback remains the same. That is - call gtk_show_uri() on the file's parent directory (probably not needed, but just in case). Also, fixed the URI generation.

Also, updated the appshell.gyp file to include the new compilation/linking flags needed for the new function to compile properly.
The library "libdbus-glib-1-dev" is now a dependency, so it should be added to the wiki page.