Fix for [CEF 2171] Crash on quit(https://github.com/adobe/brackets/issues/7683)
This is to fix the crash on quit(Issue#7683) after
integrating CEF 2171.
On Destruction, callback_map_ was getting destroyed
in the ClientApp::~ClientApp(). However while removing
all the elements, it was trying to destroy some stale
objects which were already deleted. So to fix this, we
are now explicitly clearing the map in ClientApp::OnContextReleased()
which gets triggered by CEF, just before releasing V8 context